+++ src/sockets.h
@@ -23,9 +23,9 @@ void socket_log_last_error(Log *log, const Socket *socket, int ret);
 bool socket_read_again(Socket *socket, int ret);
 
 void socket_destroy(Socket *socket);
-ssize_t socket_write(Socket *socket, const void *buffer, size_t buffer_size);
-ssize_t socket_write_cstr(Socket *socket, const char *cstr);
-ssize_t socket_write_sv(Socket *socket, String_View sv);
-ssize_t socket_read(Socket *socket, void *buffer, size_t buffer_size);
+int socket_write(Socket *socket, const void *buffer, size_t buffer_size);
+int socket_write_cstr(Socket *socket, const char *cstr);
+int socket_write_sv(Socket *socket, String_View sv);
+int socket_read(Socket *socket, void *buffer, size_t buffer_size);
 
 #endif // SOCKETS_H_
